Implement scheduler
Scheduling seems to be working fine, after fixing two bugs that
were extremely challenging to find. They were:

  - Assuming %ebp will always have a valid value in spinlock.getpcs:
    This assumption no longer holds in user mode, where the initial
    %ebp value will be zero. I guess I could have set it to
    0xFFFF_FFFF like I'm doing on kernel startup.
  - For reasons that elude me so far, string.memmove was copying
    garbage to the init process page holding the code, so I was
    getting an ILLOP trap when switching to the init process.
    Using @memcpy fixed this.

With these two bugs out of the way, I seem to have a working scheduler.
I assumed I'd lose my sanity debugging the hell out of the stack
switching trickery, but in the end it wasn't that bad.

And so, hopefully concludes the hardest part of the OS.

** This is MEGA! **